move Wave class to wave.py.
